point elevations fixed (tested for AW3D and SRTMGL1)

+++ DEM.Net.Core/Services/Interpolation/BilinearInterpolator.cs
@@ -49,13 +49,13 @@ namespace DEM.Net.Core.Interpolation
     /// </summary>
     public class BilinearInterpolator : IInterpolator
     {
-        public float Interpolate(float southWestHeight, float southEastHeight, float northWestHeight, float northEastHeight, float x, float y)
+        public double Interpolate(double southWestHeight, double southEastHeight, double northWestHeight, double northEastHeight, double x, double y)
         {
-            float h1 = southWestHeight, h2 = southEastHeight, h3 = northWestHeight, h4 = northEastHeight;
+            double h1 = southWestHeight, h2 = southEastHeight, h3 = northWestHeight, h4 = northEastHeight;
             // bilinear
-            float ha = MathHelper.Lerp(h3, h1, y);
-            float hb = MathHelper.Lerp(h4, h2, y);
-            float hi_linear = MathHelper.Lerp(ha, hb, x);
+            double ha = MathHelper.Lerp(h3, h1, y);
+            double hb = MathHelper.Lerp(h4, h2, y);
+            double hi_linear = MathHelper.Lerp(ha, hb, x);
 
             return hi_linear;
         }
